refactor: error-boundary를 사용하여 에러 띄우기
