+++ package.json
@@ -15,6 +15,7 @@
     "@tanstack/react-query-devtools": "^5.40.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-error-boundary": "^4.0.13",
     "soha-components": "^0.0.11",
     "styled-components": "^6.1.11"
   },
