+++ src/components/CustomErrorBoundary/CustomErrorBoundary.style.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+import { DefaultStyle } from "../ProductList/ProductList.style";
+
+export const Error = styled(DefaultStyle)`
+  height: 100vh;
+  font-size: 23px;
+  font-weight: 500;
+  text-align: center;
+  line-height: 40px;
+`;
