+++ src/components/ProductList/ProductList.tsx
@@ -9,8 +9,6 @@ const ProductList = () => {
     isLoading,
     isFetching,
     hasNextPage,
-    isError,
-    error,
     handleCategory,
     handleSort,
     fetchNextPage,
@@ -29,14 +27,6 @@ const ProductList = () => {
         handleSort={handleSort}
       />
       {(() => {
-        if (isError && error) {
-          return (
-            <PL.Error>
-              🚨 Error! 🚨 <br />
-              {error?.message}
-            </PL.Error>
-          );
-        }
         if (products.length === 0 && !isLoading) {
           return <PL.Empty>상품이 존재하지 않습니다! 🥲</PL.Empty>;
         }
