+++ src/main.tsx
@@ -5,6 +5,7 @@ import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
 import GlobalStyle from "./GlobalStyle.style.ts";
 import { ThemeProvider } from "styled-components";
 import { theme } from "./theme.ts";
+import CustomErrorBoundary from "@components/CustomErrorBoundary/CustomErrorBoundary.tsx";
 
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import React from "react";
@@ -14,6 +15,7 @@ const queryClient = new QueryClient({
     queries: {
       staleTime: 20000,
       refetchOnWindowFocus: false,
+      throwOnError: true,
     },
   },
 });
@@ -24,7 +26,9 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
       <QueryClientProvider client={queryClient}>
         <ReactQueryDevtools initialIsOpen />
         <GlobalStyle />
-        <App />
+        <CustomErrorBoundary>
+          <App />
+        </CustomErrorBoundary>
       </QueryClientProvider>
     </ThemeProvider>
     ,
